requirements.txt
joblib==1.4.2
numpy==1.26.4
requests==2.32.3
pyserial==3.5
scikit-learn==1.6.1
xgboost==2.0.3
pytest==8.3.5
urllib3<2
